refactor: LottoGenerator 클래스 삭제 및 기능 Lotto 클래스로 이전
