+++ src/main/java/lotto/Application.java
@@ -4,14 +4,14 @@ import java.util.List;
 
 import lotto.domain.Lotto;
 import lotto.domain.LottoGame;
-import lotto.domain.LottoGenerator;
 import lotto.domain.WinningLotto;
 import lotto.view.LottoView;
 
 public class Application {
+    public static final int LOTTO_PRICE = 1000;
 
     public static void main(String[] args) {
-        List<Lotto> lottos = LottoGenerator.generate(LottoView.inputInit());
+        List<Lotto> lottos = Lotto.autoGenerateLottos(LottoView.inputInit() / LOTTO_PRICE);
         LottoView.printLottoList(lottos);
         WinningLotto winningLotto = LottoView.inputWinningLotto();
         LottoGame lottoGame = new LottoGame(lottos, winningLotto);
