+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -10,7 +10,7 @@ class LottoGameTest {
     @Test
     void 로또갯수만큼_로또가_생성된다() {
         LottoGame lottoGame = new LottoGame(
-                LottoGenerator.generate(30000),
+                Lotto.autoGenerateLottos(30),
                 new WinningLotto(List.of(1, 2, 3, 4, 5, 6), 7)
         );
         assertThat(lottoGame.getLottos().size()).isEqualTo(30);
